// src/core/enums/tam.enums.ts
export enum ColorsEnum {
  Primary = 'primary',
  Secondary = 'secondary',
  Danger = 'danger',
  Success = 'success',
  Warning = 'warning',
  Accent = 'accent',
  Muted = 'muted',
  White = 'white',
  Gradual = 'Gradual',
}

export enum VariantsEnum {
  Filled = 'filled',
  Outline = 'outline',
  Muted = 'muted',
  Link = 'link',
  None = 'none',
  White = 'white',
  Gradual = 'Gradual',
}

export enum SizesEnum {
  Small = 'sm',
  Medium = 'md',
  Large = 'lg',
}